-enh more clean up, added comments
